Updating algebra from interpreter implementation
The following changes/fixes have been made after working on an
interpreter for the algebra.

  - Hiding Set from StringAlgebra
  - Explicitly using Scala's Set
  - Fixing Hincrby command names
  - Using Long for seconds
  - Removing num keys from Zinter/unionstore

+++ src/main/scala/redis/algebra/package.scala
@@ -3,7 +3,7 @@ package redis
 import scalaz.{Coproduct, Free}
 
 package object algebra {
-  type Seconds = Int
+  type Seconds = Long
 
   type Milliseconds = Long
